fix(spa): only parse complete ISO string to Date
This works around a problem where partial dates in registry data
are formatted as long ISO strings.

The downside of the workaround is that lookups by date may not work
for those fields.

Proper fix could be to store those dates with precision,
which would allow to both deserialize them to dates
and also format accordingly.